test(Customer, LottoStore): LottoStore 클래스의 generateLotto 메소드 관련 테스트를 새 파일로 분리
